Most of GUI finished and functional
Most of the GUI stuff has been finished.  All that's left to do is the
creation and implementation of node class to represent points to place
tokens (pieces) like nabil and I talked about.  Other than that, error
checking and error messaging still need to be implemented (if the pieces
have been placed correctly).